Added searching for follows by letter
If you want to lookup the follwers or following of a user alphabetically, you now can. The alphabet is displayed with links to search for each letter.
